transform stream for logger to the class and moved it to logger.js
