+++ src/logger.js
@@ -38,6 +38,25 @@ class Logger {
   }
 }
 
+class StreamForLogger {
+  constructor(logFileStream, date) {
+    this.logFileStream = logFileStream;
+    this.date = date;
+  }
+  write(msg) {
+    process.stdout.write(msg);
+    const currentDate = new Date().toISOString().substring(0, 10);
+    if (currentDate !== this.date) {
+      this.date = currentDate;
+      this.logFileStream.end();
+      const filePath = path.join(LOG_FOLDER_PATH, `${this.date}.log`);
+      this.logFileStream = fs.createWriteStream(filePath, { flags: 'a' });
+    }
+    this.logFileStream.write(msg);
+  }
+};
+
 module.exports = {
   Logger,
+  StreamForLogger,
 };
